+++ music-web/src/main/java/com/music/web/service/CollectionService.java
@@ -11,14 +11,14 @@ import java.util.List;
 public interface CollectionService {
 
     /*获取某个用户的收藏*/
-    public List<Music> selectMusicByUser(Long uid, int type);
+    public List<Music> selectMusicByUser(Integer uid, int type);
 
-    public boolean checkCollectionExist(Long mid,Long uid, int type);
+    public boolean checkCollectionExist(Long mid,Integer uid, int type);
 
     /* 进行插入收藏操做*/
     public boolean insertCollection(CollectionMusic collectionMusic);
 
     /* 进行删除操做*/
-    public boolean deleteCollection( Long mid,Long uid, int type);
+    public boolean deleteCollection( Long mid,Integer uid, int type);
 }
